// sdk/super_txn/src/generated/types/SuperTransactionMessage.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as web3 from '@solana/web3.js'
import * as beet from '@metaplex-foundation/beet'
import * as beetSolana from '@metaplex-foundation/beet-solana'
import {
  SuperCompiledInstruction,
  superCompiledInstructionBeet,
} from './SuperCompiledInstruction'
import {
  SuperMessageAddressTableLookup,
  superMessageAddressTableLookupBeet,
} from './SuperMessageAddressTableLookup'
export type SuperTransactionMessage = {
  numSigners: number
  numWritableSigners: number
  numWritableNonSigners: number
  accountKeys: web3.PublicKey[]
  instructions: SuperCompiledInstruction[]
  addressTableLookups: SuperMessageAddressTableLookup[]
}

/**
 * @category userTypes
 * @category generated
 */
export const superTransactionMessageBeet =
  new beet.FixableBeetArgsStruct<SuperTransactionMessage>(
    [
      ['numSigners', beet.u8],
      ['numWritableSigners', beet.u8],
      ['numWritableNonSigners', beet.u8],
      ['accountKeys', beet.array(beetSolana.publicKey)],
      ['instructions', beet.array(superCompiledInstructionBeet)],
      ['addressTableLookups', beet.array(superMessageAddressTableLookupBeet)],
    ],
    'SuperTransactionMessage'
  )